Remove duplicate API requests
